WIP, towards autofilling by symmetry
