+++ debug/viewer_lattice_axes_debug.py
@@ -4,7 +4,7 @@ from pyspinw.exchange import HeisenbergExchange
 from pyspinw.gui.viewer import show_object
 from pyspinw.hamiltonian import Hamiltonian
 from pyspinw.site import LatticeSite
-from pyspinw.structures import Structure
+from pyspinw.structure import Structure
 from pyspinw.symmetry.unitcell import UnitCell
 
 if __name__ == "__main__":
