+++ debug/viewer_test_nonmagnetic.py
@@ -5,7 +5,7 @@ from pyspinw.exchange import HeisenbergExchange
 from pyspinw.gui.viewer import show_object
 from pyspinw.hamiltonian import Hamiltonian
 from pyspinw.site import LatticeSite
-from pyspinw.structures import Structure
+from pyspinw.structure import Structure
 from pyspinw.symmetry.supercell import TiledSupercell
 from pyspinw.symmetry.unitcell import UnitCell
