+++ examples/tutorials/ferromagnetic_chain_series.py
@@ -6,7 +6,7 @@ from pyspinw.interface import generate_exchanges, filter
 from pyspinw.path import Path
 from pyspinw.site import LatticeSite
 from pyspinw.symmetry.unitcell import UnitCell
-from pyspinw.structures import Structure
+from pyspinw.structure import Structure
 
 
 unit_cell = UnitCell(1,1,1)
