+++ examples/tutorials/powder_fitting.py
@@ -10,7 +10,7 @@ from pyspinw.path import Path1D
 from pyspinw.sample import Powder
 from pyspinw.site import LatticeSite
 from pyspinw.symmetry.unitcell import UnitCell
-from pyspinw.structures import Structure
+from pyspinw.structure import Structure
 
 unit_cell = UnitCell(3, 8, 8)
