+++ examples/tutorials/swapping_chain_series.py
@@ -9,7 +9,7 @@ from pyspinw.path import Path
 from pyspinw.site import LatticeSite
 from pyspinw.symmetry.supercell import SummationSupercell, CommensuratePropagationVector
 from pyspinw.symmetry.unitcell import UnitCell
-from pyspinw.structures import Structure
+from pyspinw.structure import Structure
 import sys
 
 from pyspinw.debug_plot import debug_plot
