+++ pyspinw/__init__.py
@@ -7,7 +7,7 @@ magnetic structures and excitations
 
 from pyspinw.site import LatticeSite
 from pyspinw.symmetry.unitcell import UnitCell
-from pyspinw.structures import Structure
+from pyspinw.structure import Structure
 from pyspinw.anisotropy import Anisotropy, AxisMagnitudeAnisotropy
 from pyspinw.exchange import HeisenbergExchange, DiagonalExchange, XYExchange, IsingExchange, DMExchange
 from pyspinw.hamiltonian import Hamiltonian
